abstract and reuse chart solution
